Constructed top 20 elements that should be counted
Completely immense code to perform a simple task of removing the values
are not commonly used. This could of done much easier with pandas,
however, used a mix of numpy and pandas.

This is very specific code, at current moment, that will be used only
for the `Key Skill` column. Not sure if I want to optimize the code or
just move on to finding a possible ways of generalizing the code to be
used on other columns.